Extract StorageSideNavigatorItem
